back/requirements.txt
Django>=4.1.3, <4.2
django-admin-interface>=0.23.0, <0.24
djangorestframework>=3.14.0, <3.15
uWSGI>=2.0.21, <2.1